Move the LPUpdate type into a separate file
This way frontend code can import it without feeling bad.
